backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
pymongo